fix(core): Rename const -> constants | Fix import
