+++ libs/core/src/lib/form/form-message/form-message.component.ts
@@ -9,7 +9,7 @@ import {
 } from '@angular/core';
 import { CssClassBuilder } from '../../utils/interfaces/css-class-builder.interface';
 import { applyCssClass } from '../../utils/decorators/apply-css-class.decorator';
-import { CSS_CLASS_NAME, getTypeClassName } from './consts';
+import { CSS_CLASS_NAME, getTypeClassName } from './constants';
 
 export type MessageStates = 'success' | 'error' | 'warning' | 'information';
